refactor(view): 입력을 파싱하는 역할을 서비스 객체에게 위임함
기존에 입력을 파싱하던 역할을 InputView에서 하는 것은
적절하지 않다고 판단하여 service단에 파싱만을 위한 객체 생성함
