+++ docs/README.md
@@ -116,12 +116,13 @@
 ---
 
 ## 📋 구현 기능 목록
-### <u>입력</u> (대부분의 예외처리는 적절한 객체가 수행합니다.)
+### <u>입력 관련</u> (대부분의 예외처리는 적절한 객체가 수행합니다.)
 - [x] 로또 구입 금액을 1000원 단위로 입력 받는다.
   - [x] `예외처리` 널 또는 공백으로만 이루어져 있는지 검사한다.
   - [x] `예외처리` 0보다 작거나 같은 값일 경우 예외를 발생시킨다.
   - [x] `예외처리` 1000원으로 나누어 떨어지지 않는 경우 예외를 발생시킨다.
   - [x] `예외처리` 정수가 아닌 경우 예외를 발생시킨다.
+  - [x] 입력 받은 문자열을 구입 금액으로 파싱한다.
 - [x] 쉼표(,)를 기준으로 당첨 번호 6자리를 입력 받는다.
   - [x] `예외처리` 널 또는 공백으로만 이루어져 있는지 검사한다.
   - [x] `예외처리` 쉼표(,)를 기준으로 나누었을 때 
@@ -129,11 +130,13 @@
     - [x] `예외처리` 입력이 6자리가 아닌 경우 예외를 발생시킨다.
   - [x] `예외처리` 1~45 사이의 숫자가 아닌 경우 예외를 발생시킨다.
   - [x] `예외처리` 중복된 숫자가 포함되어 있을 경우 예외를 발생시킨다.
+  - [x] 입력 받은 문자열을 당첨 로또로 파싱한다.
 - [x] 보너스 번호를 입력 받는다.
   - [x] `예외처리` 널 또는 공백으로만 이루어져 있는지 검사한다.
   - [x] `예외처리` 숫자가 아닌 경우 예외를 발생시킨다.
   - [x] `예외처리` 1~45 사이의 숫자가 아닌 경우 예외를 발생시킨다.
   - [x] `예외처리` 당첨 번호와 중복이 있을 경우 예외를 발생시킨다.
+  - [x] 입력 받은 문자열을 보너스 번호로 파싱한다.
 
 ### <u>핵심 로직</u>
 - [x] 로또 당첨 번호를 생성한다.
@@ -148,14 +151,14 @@
       - [x] 출력 포메팅에 맞게 로또 번호들을 배치한다.
 - [x] 발행된 로또와 당첨 번호를 비교하여 추첨을 진행한다. 
   - [x] 발행된 로또의 개수가 여러 개인 경우 반복하여 추첨을 진행한다. 
-  - [x] 로또 번호와 당첨[Lottos.java](..%2Fsrc%2Fmain%2Fjava%2Flotto%2Fdomain%2FLottos.java) 번호를 비교해 일치하는 개수를 알아낸다.
+  - [x] 로또 번호와 당첨 번호를 비교해 일치하는 개수를 알아낸다.
   - [x] 로또 번호가 보너스 번호를 포함하고 있는지 알아낸다.
   - [x] 일치하는 개수에 맞춰 등수를 계산한다. 
 - [x] 구입 금액과 당첨 금액의 비율로 수익률을 계산한다.
   - [x] 당첨 현황을 보고 당첨 금액을 계산한다.
 - [x] 로또 게임을 컨트롤한다.
   - [x] 입출력 값을 연결하여 게임을 진행한다.
-### <u>출력</u> 
+### <u>출력 관련</u> 
 - [x] 구입 금액 입력 요청 메시지를 출력한다.
 - [x] 구매한 로또 개수를 출력한다.
 - [x] 발행된 로또를 전부 출력한다.
